Feature/raport module

+++ autopartsJee/src/main/java/pl/isa/autopartsJee/carToDatabase/dao/CarRepositoryDaoBean.java
@@ -2,7 +2,6 @@ package pl.isa.autopartsJee.carToDatabase.dao;
 
 import pl.isa.autopartsJee.carToDatabase.domain.CarData;
 import pl.isa.autopartsJee.carToDatabase.repository.CarRepository;
-
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import java.util.List;
@@ -18,12 +17,12 @@ public class CarRepositoryDaoBean implements CarRepositoryDao {
     }
 
     @Override
-    public List<CarData> findCarsByOwnerId(int ownerId) {
+    public List<CarData> findCarsByOwnerId(Long ownerId) {
         return carRepository.findCarsByOwnerId(ownerId);
     }
 
     @Override
-    public CarData findCarById(Integer carID) {
+    public CarData findCarById(Long carID) {
         return carRepository.findCarById(carID);
     }
 }
